+++ ligo/skymap/tool/ligo_skymap_contour_moc.py
@@ -1,0 +1,78 @@
+#
+# Copyright (C) 2013-2020 Giuseppe Greco, Leo Singer and CDS team.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+"""
+Create a contour for the credible levels of an all-sky probability
+in a standard MOC (Multi Order Coverage) data structure
+(http://ivoa.net/documents/MOC/).
+The input can be either a multiresolution or a flattened HEALPix
+probability map. The output is stored in a fits format.
+"""
+
+from . import ArgumentParser, FileType
+
+
+def parser():
+    parser = ArgumentParser()
+
+    parser.add_argument(
+        '--output', metavar='FILE.fits',
+        default='-', type=str, help='output file [default: stdout]')
+    parser.add_argument(
+        '-c', '--contour', metavar='PERCENT', type=float, required=True,
+        help='MOC region enclosing this percentage of probability \
+              [range is 0-100]')
+    parser.add_argument(
+        'input', metavar='INPUT.fits[.gz]', type=FileType('rb'),
+        default='-', nargs='?', help='Input multi-order or flatten \
+                                      HEALPix FITS file')
+
+    return parser
+
+
+def main(args=None):
+    pars = parser()
+    opts = parser().parse_args(args)
+
+    import astropy_healpix as ah
+    import astropy.units as u
+
+    try:
+        from mocpy import MOC
+    except ImportError:
+        pars.error('This command-line tool requires mocpy >= 0.8.2.'
+                   'Please install it by running "pip install mocpy".')
+
+    from ..io import read_sky_map
+
+    # Read multi-order sky map
+    skymap = read_sky_map(opts.input.name, moc=True)
+
+    uniq = skymap['UNIQ']
+    probdensity = skymap['PROBDENSITY']
+
+    level, ipix = ah.uniq_to_level_ipix(uniq)
+    area = ah.nside_to_pixel_area(
+        ah.level_to_nside(level)).to_value(u.steradian)
+
+    prob = probdensity * area
+
+    # Create MOC
+    contour_decimal = opts.contour / 100
+    moc = MOC.from_valued_healpix_cells(
+        uniq, prob, cumul_from=0.0, cumul_to=contour_decimal)
+
+    # Write MOC
+    moc.write(opts.output, format='fits', overwrite=True)